Remove unused net/rpc based plugin implementations
