+++ go.mod
@@ -6,9 +6,7 @@ require (
 	github.com/google/go-cmp v0.5.7
 	github.com/hashicorp/go-hclog v1.2.0
 	github.com/hashicorp/go-plugin v1.4.3
-	github.com/hashicorp/go-version v1.4.0
 	github.com/hashicorp/hcl/v2 v2.11.1
-	github.com/hashicorp/terraform-svchost v0.0.0-20200729002733-f050f53b9734
 	github.com/zclconf/go-cty v1.10.0
 	google.golang.org/grpc v1.45.0
 	google.golang.org/protobuf v1.27.1
