--- tflint/server/doc.go
@@ -1,2 +1,0 @@
-// Package server contains the interfaces that the host process should satisfy.
-package server
